fix(scribe): 突き合わせの両側を再び別ファイルへ分ける
前のコミットで structure_page.py をテストへ畳んだが、テスト側は同じ名前の
_frontmatter_lines と _section_body を独立実装として持っており、後から定義された
テスト側が畳んだ側を shadow していた。find_structure_pages が解決するのも
テスト側の実装になり、「独立に導出して突き合わせる」検査が自分自身との比較に
なっていた (TESTING.md の「成果物から入力を組む」に当たる)。

module を戻して両側を分ける。指摘のあった呼び出し元の無い CLI (main と argv 解析)
は落とし、`.ja` ミラーを置く。英語側の docstring に残っていた節名は data として
引用する。
